fixing wait_connection_availability()
wait_connection_availability() did take at least 2 seconds per call in
the old incarnation. Now we may finish a call without any sleep at all
when the result is already ready at the time called

+++ dbutils.h
@@ -39,6 +39,6 @@ const char	 *get_cluster_size(PGconn *conn);
 PGconn *getMasterConnection(PGconn *standby_conn, char *schema, char *cluster,
                             int *master_id, char *master_conninfo_out);
 
-int wait_connection_availability(PGconn *conn, int timeout);
+int wait_connection_availability(PGconn *conn, unsigned long timeout);
 bool CancelQuery(PGconn *conn, int timeout);
 #endif
